+++ River/River/RiverUI/Header/UIWindow.h
@@ -14,7 +14,7 @@ public:
 
 	void OnUpdate(float deltaTime);
 
-	void OnRender(V_Array<UIVertex>& vertices, V_Array<uint32_t>& indices);
+	void OnRender(V_Array<UIVertex>& vertices, V_Array<uint16_t>& indices);
 
 	void AddPanel(int orderLevel, Unique<Panel>& panel);
