+++ River/River/RiverUI/Source/Widget.cpp
@@ -16,7 +16,7 @@ void Widget::OnUpdate(float deltaTime)
 
 }
 
-void Widget::OnRender(V_Array<UIVertex>& vertices, V_Array<uint32_t>& indices)
+void Widget::OnRender(V_Array<UIVertex>& vertices, V_Array<uint16_t>& indices)
 {
 	m_RenderItem->OnRender(vertices, indices);
 }
